Handle windows terminal characteristics
Specifically git rid of warnings for the windows VS compiler that
doesn't allow the buffer to be allocated dynamically on the stack
and change it to malloc and free.

Also handle the cursor positioning so that it doesn't go to the
next line unless there is actually a character past the end of the
buffer, since the windows terminal will send a CRCRLF sequence
when the end of the line is reached, and the writecharatcurs method
would have already inserted a line.  Also the Windows CMD.EXE tries
to write out a screen title by doing an OS command using <ESC>]0;
sequence which the handle_char now silently consumes.