Change CSI n X to start at cursor position and go n characters
The current processing for CSI n X started at the cursor position but ended
based on the 'n' from the beginning of the line.  It should have been
from the current cursor position.  This shows up very readily when
handling powershell commands that provide hints and then erase the
hint when the user backspaces to the beginning of the prompt (where
the prompt is often quite long).

This change adjusts the CSI n X to go from the cursor to n + cursor position.