Add compile time feature to force use of UTF8
The current implementation uses the wchar_t and multi-byte to
wide char functions to handle characters that don't fit in a single
byte.  Unfortunately this doesn't work well in Windows which has
default multi-byte handling in UTF-16 which requires 2 wchar_t to
represent a single glyph in the U+10FFFF range.

Also all communication using the ConPTY
for Windows is in UTF-8 so that is another inconsistency that would
require converting to and from UTF-8, UTF-16, and UTF-32.

Another issue is that just because the platform has a multi-byte encoding
one way doesn't mean that the input from a remote system might not
be something else.  The vast majority of systems however will do
UTF-8, and it is not unreasonable for some applications to restrict
themselves to working with UTF-8 systems.

This change includes a compilation flag that forces the libtmt library
to interpret the character streams as containing UTF-8 characters
instead of relying on the locale, and has it storing glyphs as 32-bit
integers which can encode all unicode characters.  It provides a simple
UTF-8 to 32-bit wide character defined
as tmt_wchar_t.  It also provides a conversion back from the wide
character to UTF-8.