Added half and full char width for unicode characters
There are a number of characters that can be what is called
fullwidth, but the character actually takes 2 columns.  This
has implecations in cursor placement and delete handling
because the application the terminal is talking to will
send 2 deletes (one for each column) even though there is
only 1 character.

This required the addition of a lookup table to determine what
characters are full width and which ones aren't.  This is then
communicated through the callbacks allowing the renderer to
adjust accordingly.