Support combining marks and formatters
Unicode supports characters that are combining marks (like
accents) and this could be in files that are output from a command
running on the other end of the terminal.

To preserve the data so it shows up to the renderer, those characters
need to be provided even though they do not take up any additional
columns in the output.

To determine if it is a mark (which occurs after a code) or
a formatter (which occurs before a code) the unicode table with
these types needs to be turned into a lookup and used by this code.

These codes are not implemented well in other terminals like
the Gnome terminal and in applications like vi, so using these
characters is not particularly reliable.

The goal for this library is that the characters are preserved so
the renderer can do it's best.  In other words if something comes
out or goes into an application running somewhere
this library doesn't lose the information that the renderer might
need.