Minimal support for cursor style
On windows the terminal codes are sent independent of a TERM variable
setting so it is more important to cover "typical" codes that are being
sent.  In this case the CSI Pn SP q sequence defines a cursor style
which can be block or underline.  This code is now handled instead of
causing the state to reset and an extra "space q" being sent out
whenever an application like "vim.exe" would send a change.

It currently simply stores the cursor style in the VT data structure but
doesn't actually pass it in the callback.  A future change could be
done to support that.