Support indexed colors
Add support for indexed colors and handle control characters and
last column.  The indexed colors are used on windows 11 by programs
such as vim.

Also since the cursor is past the last column (in case of controls like
backspace or cursor back on the same line) and the rolling to the next
line doesn't occur until another character is written the fixcursor
method cannot reset it back to the position prior to the end.